Add files via upload
This project comprises of a guests' name stored in an array. If new guest inputs name and name is found in the array, DOM is used to manipulate the heading text to welcome guest along with their name mentioned, else if guest name isn't included in the array of expected guests, guest is denied access.