My 8th JavaScript commit
The to do list project allows a user to type a task say "Practice coding" into an input field. This task is then captured and displayed in paragraphs using JavaScript. On completion of a task, the user can click on the task to cancel it or double click it to remove the task from the listed paragraphs.